+++ CHANGELOG.md
@@ -8,3 +8,5 @@
    * ParetoXX example now works.
 2. Added error values to regression function.
 3. Updated formatting in manual.
+4. Added n arguments to d2cor and cor2d.
+5. Added Jeff as author.